Struct mccs_db::Database

pub struct Database { /* fields omitted */ }

Describes all the VCP feature codes supported by an MCCS specification or display.

Methods

impl Database

pub fn from_version(mccs_version: &Version) -> Self

Create a new database from a specified MCCS specification version.

pub fn from_database<R: Read>(
database_yaml: R,
mccs_version: &Version
) -> Result<Self>

Create a new database from a specified database description YAML file.

This format is not (yet) documented, but an example exists that describes the MCCS spec.

pub fn apply_capabilities(&mut self, caps: &Capabilities)

Filter out any feature codes or values that are not supported by the specified display.

pub fn get(&self, code: FeatureCode) -> Option<&Descriptor>

Get the description of a given VCP feature code.