Struct mccs_db::Descriptor

pub struct Descriptor {
    pub name: Option<String>,
    pub description: Option<String>,
    pub group: Option<String>,
    pub code: FeatureCode,
    pub ty: ValueType,
    pub access: Access,
    pub mandatory: bool,
    pub interacts_with: Vec<FeatureCode>,
}

Describes a VCP feature code's functionality and value format.

Fields
name: Option<String>

The name of the feature.

description: Option<String>

A detailed description of the feature.

group: Option<String>

The MCCS grouping this feature belongs to.

code: FeatureCode

The VCP code of the feature.

ty: ValueType

The data type of the feature.

access: Access

Whether the feature can be set, read, or both.

mandatory: bool

Whether the feature is required to be supported by the display for MCCS specification compliance.

interacts_with: Vec<FeatureCode>

Any other feature codes that this "interacts" with.

Changing this feature's value may also affect the value of these other listed features.